requirements.txt
discord.py
requests
nekos.py
